% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex1.1}
\alias{ex1.1}
\title{Data for Example 1.1}
\format{A data frame with 50 observations of 3 variables:
\describe{
  \item{State}{}
  \item{Internet}{}
  \item{College}{}
}}
\usage{
ex1.1
}
\description{
[Some description here]
}
\keyword{datasets}
