% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex10.12}
\alias{ex10.12}
\title{Data for Example 10.12}
\format{A data frame with 6 observations of 5 variables:
\describe{
  \item{Condition}{}
  \item{Criticism}{}
  \item{Symptoms}{}
}}
\usage{
ex10.12
}
\description{
[Some description here]
}
\keyword{datasets}
