% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex11.1}
\alias{ex11.1}
\title{Data for Example 11.1}
\format{A data frame with 6 observations of 5 variables:
\describe{
  \item{group}{}
  \item{instructions}{}
  \item{score1}{}
  \item{score2}{}
  \item{score3}{}
}}
\usage{
ex11.1
}
\description{
[Some description here]
}
\keyword{datasets}
