% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex11.22}
\alias{ex11.22}
\title{Data for Example 11.22}
\format{A data frame with 16 observations of 5 variables:
\describe{
  \item{couple}{}
  \item{children}{Does the couple have children? ys or no}
  \item{yearsmarried}{Number of years married}
  \item{Male}{}
  \item{Female}{}
}}
\usage{
ex11.22
}
\description{
Scores for husbands and wives by couple with years married and child status.
}
\details{
TODO: What are the couple, Male and Female scores?
}
\keyword{datasets}
