% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex11.9}
\alias{ex11.9}
\title{Exhibit 11.9}
\format{A data frame with 8 observations of 3 variables.
\describe{
  \item{subject}{Subject ID}
  \item{absent}{Performance with experimenter absent (1 - 9, 9 is high)}
  \item{present}{Performance with experimenter present (1 - 9, 9 is high)}
}}
\source{
<http://www.dataanalysisbook.com/datasetsInTextbook.html>
}
\usage{
ex11.9
}
\description{
Data from a crossed design where participants completed a set of math puzzles
with the experimenter in the room and a set when the experimenter was absent.
}
\keyword{datasets}
