% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex5.1}
\alias{ex5.1}
\title{Exhibit 5.1}
\format{A data frame with 50 observations of 3 variables:
\describe{
  \item{State}{}
  \item{Internet}{}
  \item{College}{}
}}
\usage{
ex5.1
}
\description{
Internet access rates and college education rates for each of the 50 U.S. states.
}
\details{
TODO: what are the units of the rates?
}
\keyword{datasets}
