% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex6.2}
\alias{ex6.2}
\title{Data for Example 6.2}
\format{A data frame with 19 observations of 5 variables:
\describe{
  \item{Name}{}
  \item{Sex}{}
  \item{Age}{}
  \item{Height}{}
  \item{Weight}{}
}}
\usage{
ex6.2
}
\description{
[Some description here]
}
\keyword{datasets}
