% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex7.1}
\alias{ex7.1}
\title{Data for Example 7.1}
\format{A data frame with 80 observations of 3 variables:
\describe{
  \item{TIME}{}
  \item{AGE}{}
  \item{MILES}{}
}}
\usage{
ex7.1
}
\description{
[Some description here]
}
\keyword{datasets}
