% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex7.2}
\alias{ex7.2}
\title{Exhibit 7.2}
\format{A data frame with 80 observations of 3 variables:
\describe{
  \item{TIME}{Amount of time spent running per week}
  \item{AGE}{Age of the runner}
  \item{MILES}{Number of miles run per week}
}}
\usage{
ex7.2
}
\description{
Running time, age, and miles of training per week for hypothetical runners.
}
\keyword{datasets}
