% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex8.1}
\alias{ex8.1}
\title{Data for Example 8.1}
\format{A data frame with 20 observations of 3 variables:
\describe{
  \item{Student}{}
  \item{Group}{}
  \item{SAT}{}
}}
\usage{
ex8.1
}
\description{
[Some description here]
}
\keyword{datasets}
