% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex8.13}
\alias{ex8.13}
\title{Data for Example 8.13}
\format{A data frame with 18 observations of 2 variables:
\describe{
  \item{Program}{}
  \item{GRE}{}
}}
\usage{
ex8.13
}
\description{
[Some description here]
}
\keyword{datasets}
