% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex9.1}
\alias{ex9.1}
\title{Data for Example 9.1}
\format{A data frame with 50 observations of 3 variables:
\describe{
  \item{Drug}{}
  \item{Psychotherapy}{}
  \item{Mood}{}
}}
\usage{
ex9.1
}
\description{
[Some description here]
}
\keyword{datasets}
