% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{US_internet_grad}
\alias{US_internet_grad}
\alias{ex1.1}
\alias{ex2.9}
\alias{ex5.1}
\title{US internet and college graduation rates}
\format{A data frame with 50 observations of 3 variables.
\describe{
  \item{State}{The two-letter abbreviation of the state.}
  \item{Internet}{The percentage of households that had internet access.}
  \item{College}{The percentage of residents with a college degree.}
}}
\source{
<http://www.dataanalysisbook.com/datasetsInTextbook.html>
}
\usage{
US_internet_grad

ex1.1

ex2.9

ex5.1
}
\description{
Internet access and college graduation rates for each of the 50 U.S. states
in the year 2000.
}
\keyword{datasets}
