% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex1.1}
\alias{ex1.1}
\title{Data for Example 1.1}
\format{A data frame with 50 observations of 3 variables:
\describe{
  \item{State}{The two-letter abbreviation of the state.}
  \item{Internet}{}
  \item{College}{}
}}
\usage{
ex1.1
}
\description{
Internet access rates for each of the 50 U.S. states.
}
\details{
TODO: what are the units of the rates?
}
\keyword{datasets}
