% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex10.1}
\alias{ex10.1}
\title{Data for Example 10.1}
\format{A data frame with 40 observations of 4 variables:
\describe{
  \item{score}{}
  \item{curriculum}{}
  \item{teacher}{}
  \item{pretest}{}
}}
\usage{
ex10.1
}
\description{
[Some description here]
}
\keyword{datasets}
