% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex10.12}
\alias{ex10.12}
\title{Exhibit 10.12}
\format{A data frame with 6 observations of 5 variables:
\describe{
  \item{Condition}{Whether the participant was in the treatment or control group}
  \item{Criticism}{Number of criticisms}
  \item{Symptoms}{Number of symptoms}
}}
\usage{
ex10.12
}
\description{
Psychological symptoms and number of criticisms across two conditions.
}
\keyword{datasets}
