% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex10.8}
\alias{ex10.8}
\title{Data for Example 10.8}
\format{A data frame with 40 observations of 4 variables:
\describe{
  \item{score}{The student's achievement score}
  \item{curriculum}{-1 = old, 1 = new}
  \item{teacher}{-1 = A, 1 = B}
  \item{pretest}{The student's score on a pretest}
}}
\usage{
ex10.8
}
\description{
Modified version of ex10.1.txt; see textbook for explanation.
}
\details{
TODO: update explanation
}
\keyword{datasets}
