% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex11.1}
\alias{ex11.1}
\title{Data for Example 11.1}
\format{A data frame with 6 observations of 5 variables:
\describe{
  \item{group}{Group ID}
  \item{instructions}{Whether the participants were instructed they were in \code{competition} or not (\code{none})}
  \item{score1}{Score on first trial}
  \item{score2}{Score on second trial}
  \item{score3}{Score on third trial}
}}
\usage{
ex11.1
}
\description{
Three within-group scores for two levels of competition.
}
\keyword{datasets}
