% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex11.17}
\alias{ex11.17}
\title{Data for Example 11.17}
\format{A data frame with 5 observations of 7 variables:
\describe{
  \item{Subject}{}
  \item{abstract1}{}
  \item{concrete1}{}
  \item{abstract2}{}
  \item{concrete2}{}
  \item{abstract3}{}
  \item{concrete3}{}
}}
\usage{
ex11.17
}
\description{
[Some description here]
}
\keyword{datasets}
