% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex11.9}
\alias{ex11.9}
\title{Data for Example 11.9}
\format{A data frame with 8 observations of 3 variables:
\describe{
  \item{subject}{}
  \item{absent}{}
  \item{present}{}
}}
\usage{
ex11.9
}
\description{
[Some description here]
}
\keyword{datasets}
