% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex11.9}
\alias{ex11.9}
\title{Exhibit 11.9}
\format{A data frame with 8 observations of 3 variables:
\describe{
  \item{subject}{Subject ID}
  \item{absent}{}
  \item{present}{}
}}
\usage{
ex11.9
}
\description{
Data for crossed design (experimenter present or absent)
}
\details{
TODO: What do the scores represent?
}
\keyword{datasets}
