% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex13.1}
\alias{ex13.1}
\title{Data for Example 13.1}
\format{A data frame with 13 observations of 5 variables:
\describe{
  \item{student}{Student ID}
  \item{sat}{Student's SAT score}
  \item{hsrank}{Student's rank in high school}
  \item{satr}{Reversed values of \code{sat}}
  \item{hsrankr}{Reversed values of \code{hsrankr}}
}}
\usage{
ex13.1
}
\description{
SAT and high school ranks with and without an outlier induced by reversing variable values for one observations.
}
\details{
TODO: these don't appear to be reversed. Check textbook for description and method
}
\keyword{datasets}
