% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex4.1}
\alias{ex4.1}
\title{Data for Example 4.1}
\format{A data frame with 20 observations of 1 variable:
\describe{
  \item{bids}{}
}}
\usage{
ex4.1
}
\description{
[Some description here]
}
\keyword{datasets}
