% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex8.1}
\alias{ex8.1}
\title{Exhibit 8.1}
\format{A data frame with 20 observations of 3 variables:
\describe{
  \item{Student}{Student ID}
  \item{Group}{Whether the students took the \code{course} or not \code{noCourse}}
  \item{SAT}{Score on the SAT}
}}
\usage{
ex8.1
}
\description{
SAT scores for two groups of student who did or did not take a prep course.
}
\keyword{datasets}
