% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex8.15}
\alias{ex8.15}
\title{Data for Example 8.15}
\format{A data frame with 15 observations of 2 variables:
\describe{
  \item{Grade}{}
  \item{Score}{}
}}
\usage{
ex8.15
}
\description{
[Some description here]
}
\keyword{datasets}
