% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex8.6}
\alias{ex8.6}
\title{Exhibit 8.6}
\format{A data frame with 24 observations of 2 variables.
\describe{
  \item{Condition}{Feeback given: failure, none, success}
  \item{Task}{Number of tasks completed}
}}
\source{
<http://www.dataanalysisbook.com/datasetsInTextbook.html>
}
\usage{
ex8.6
}
\description{
Hypothetical experimental data for a motivation study where children perform
a task, are given feedback, and then monitored for the number of tasks they
subsequently complete. In one condition children were given feedback that
they failed the task, other children were given no feedback, and in the last
condition children were told they were successful.
}
\keyword{datasets}
