% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex8.6}
\alias{ex8.6}
\title{Data for Example 8.6}
\format{A data frame with 24 observations of 2 variables:
\describe{
  \item{Condition}{Feeback given: failure, none, success}
  \item{Task}{Score on the task}
}}
\usage{
ex8.6
}
\description{
Subsequent tasks completed after different types of feedback on a previous task.
}
\keyword{datasets}
