% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex9.1}
\alias{ex9.1}
\title{Exhibit 9.1}
\format{A data frame with 50 observations of 3 variables.
\describe{
  \item{Drug}{The drug given: DrugA, DrugB, or Placebo}
  \item{Psychotherapy}{Psychotherapy condition: Treatment or Control}
  \item{Mood}{Patient's mood score}
}}
\source{
<http://www.dataanalysisbook.com/datasetsInTextbook.html>
}
\usage{
ex9.1
}
\description{
Hypothetical mood scores in a Drug (3) x Psychotherapy (2) factorial design.
}
\keyword{datasets}
