% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex9.1}
\alias{ex9.1}
\title{Exhibit 9.1}
\format{A data frame with 50 observations of 3 variables:
\describe{
  \item{Drug}{The drug given: A, B, or Placebo}
  \item{Psychotherapy}{Type of psychotherapy administered: treatment or control}
  \item{Mood}{Patient's mood score}
}}
\usage{
ex9.1
}
\description{
Mood scores from a drug x psychotherapy (treatment, control) factorial design.
}
\keyword{datasets}
