% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lottery}
\alias{lottery}
\title{Lottery dataset}
\format{A data frame with 20 observations of 1 variable:
\describe{
  \item{bids}{}
}}
\usage{
lottery
}
\description{
[Some description here]
}
\keyword{datasets}
